Remove TensorBase
Not subclassed except by Tensor. Also requried to align further with
caffe2.
